🚛 Extract Spotify.Credentials
