+++ lib/speed/spotify/client.ex
@@ -23,43 +23,16 @@ defmodule Speed.Spotify.Client do
   ## Next step
   curl -H "Authorization: Basic <client_n_secret>" -d grant_type=authorization_code -d code=<code> -d redirect_uri=http://localhost/callback https://accounts.spotify.com/api/token
   """
-  alias Speed.Spotify.SpotifyBehavior
+  alias Speed.Spotify.{SpotifyBehavior, Credentials}
 
   use Agent
   @behaviour SpotifyBehavior
 
-  @type t :: %__MODULE__{
-          access_token: String.t(),
-          client_id: String.t(),
-          client_secret: String.t(),
-          refresh_token: String.t()
-        }
-
-  defstruct [
-    :access_token,
-    :client_id,
-    :client_secret,
-    :refresh_token
-  ]
-
-  @spec refresh_access_token(t()) :: t()
-  def refresh_access_token(credentials) do
-    case request_refreshed_access_token(credentials) |> IO.inspect(label: "refresh_access_token") do
-      %Req.Response{body: %{"access_token" => at}} ->
-        %__MODULE__{credentials | access_token: at}
-
-      _anything_else ->
-        raise "AHHH"
-    end
-  end
-
-  def base_url, do: "https://api.spotify.com/v1/me/top/"
-
   @impl SpotifyBehavior
-  def top(creds \\ get_creds()) do
+  def top(creds \\ Credentials.get_creds()) do
     # IO.inspect(creds, label: "top with these creds")
     url = base_url() <> "artists"
-    headers = to_get_headers(creds)
+    headers = Credentials.to_get_headers(creds)
 
     case Req.get!(url, headers: headers) do
       %Req.Response{status: 200, body: %{"items" => artists}} ->
@@ -68,7 +41,7 @@ defmodule Speed.Spotify.Client do
       %Req.Response{body: %{"error" => %{"message" => "The access token expired", "status" => 401}}} ->
         creds
         |> refresh_access_token()
-        |> put_creds()
+        |> Credentials.put_creds()
         |> top()
 
       {:error, error} ->
@@ -76,60 +49,26 @@ defmodule Speed.Spotify.Client do
     end
   end
 
-  # TODO - replace by writing to the DB
-  @doc "The `Agent` is started with an empty `Credentials` struct"
-  def start_link(_) do
-    Agent.start_link(fn -> nil end, name: CredStore)
-  end
+  def base_url, do: "https://api.spotify.com/v1/me/top/"
 
-  @spec get_creds :: t()
-  def get_creds do
-    case Agent.get(CredStore, & &1) do
-      nil -> default_creds()
-      creds -> creds
+  @spec refresh_access_token(Credentials.t()) :: Credentials.t()
+  defp refresh_access_token(credentials) do
+    case request_refreshed_access_token(credentials) |> IO.inspect(label: "refresh_access_token") do
+      %Req.Response{body: %{"access_token" => at}} ->
+        %Credentials{credentials | access_token: at}
+
+      _anything_else ->
+        raise "AHHH"
     end
   end
 
-  defp put_creds(creds) do
-    Agent.update(CredStore, fn _ -> creds end)
-    creds
-  end
-
-  @spec default_creds :: t()
-  defp default_creds do
-    %__MODULE__{
-      access_token: Application.fetch_env!(:speed, Speed.Spotify)[:access_token],
-      client_id: Application.fetch_env!(:speed, Speed.Spotify)[:client_id],
-      client_secret: Application.fetch_env!(:speed, Speed.Spotify)[:client_secret],
-      refresh_token: Application.fetch_env!(:speed, Speed.Spotify)[:refresh_token]
-    }
-  end
-
-  def refresh_creds do
-    creds = get_creds()
-    request_refreshed_access_token(creds)
-  end
-
-  @spec request_refreshed_access_token(t()) :: Req.Response.t()
+  @spec request_refreshed_access_token(Credentials.t()) :: Req.Response.Credentials.t()
   def request_refreshed_access_token(%{refresh_token: rt} = creds) do
     endpoint = "https://accounts.spotify.com/api/token"
     body = "grant_type=refresh_token&refresh_token=#{rt}"
 
-    headers = to_refresh_headers(creds)
+    headers = Credentials.to_refresh_headers(creds)
 
     Req.post!(endpoint, body: body, headers: headers)
   end
-
-  @spec to_refresh_headers(t()) :: [{String.t(), String.t()}]
-  def to_refresh_headers(%{client_id: ci, client_secret: cs}) do
-    encoded_credentials = :base64.encode("#{ci}:#{cs}")
-
-    [
-      {"Authorization", "Basic #{encoded_credentials}"},
-      {"Content-Type", "application/x-www-form-urlencoded"}
-    ]
-  end
-
-  @spec to_get_headers(t()) :: [{String.t(), String.t()}]
-  def to_get_headers(%{access_token: at}), do: [{"Authorization", "Bearer #{at}"}]
 end
